computer_graphics/lab3/pyproject.toml
[tool.poetry]
name = "lab2"
version = "0.1.0"
description = ""
authors = ["Vitaly Kopylov <workflow.elec@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyopengl = "^3.1.6"
pyopengl-accelerate = "^3.1.6"
numpy = "^1.24.2"
pillow = "^9.4.0"
pygame = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
